login functionality cleaned up
