+++ @lingosta/app/app/auth/login/page.tsx
@@ -16,23 +16,25 @@ const login = () => {
 
   return (
     <div className='mx-auto p-5'>
-      <form onSubmit={handleLogin} className=''>
-        <input
-          type='email'
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-          placeholder='Email'
-          className='text-black'
-        />
-        <input
-          type='password'
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          placeholder='Password'
-          className='text-black'
-        />
-        <button type='submit'>Login</button>
-      </form>
+      <div>
+        <form onSubmit={handleLogin} className=''>
+          <input
+            type='email'
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            placeholder='Email'
+            className='text-black'
+          />
+          <input
+            type='password'
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            placeholder='Password'
+            className='text-black'
+          />
+          <button type='submit'>Login</button>
+        </form>
+        </div>
     </div>
   )
 }
