+++ @lingosta/app/app/layout.tsx
@@ -1,5 +1,5 @@
-import Layout from '../components/Layout'
-import Navbar from '../components/Navbar'
+import Layout from '../components/Layout/Layout'
+import Navbar from '../components/Layout/Navbar'
 import './globals.css'
 import { Inter } from 'next/font/google'
