+++ @lingosta/app/app/page.tsx
@@ -1,10 +1,18 @@
+'use client'
 import useUser from "../hooks/useUser"
 
 export default function Home() {
 
+  const { user } = useUser();
+
+
   return (
     <div>
-      Lingosta
+      {user && (
+        <div>
+          hello {user.name}
+        </div>
+      )}
     </div>
   )
 }
