+++ @lingosta/app/components/Layout/Layout.tsx
@@ -1,16 +1,10 @@
 'use client'
 
 import React, { useEffect } from 'react'
-import useUser from '../hooks/useUser'
+import useUser from '../../hooks/useUser'
 import { useRouter } from 'next/navigation';
 
 const Layout = ({ children }) => {
-
-    const { checkUser } = useUser();
-
-    useEffect(() => {
-        checkUser()
-    },[checkUser])
 
     return (
     <div className='container h-full mx-auto'>
