--- @lingosta/app/components/Navbar.tsx
@@ -1,27 +1,0 @@
-'use client';
-
-import React from 'react'
-import useUser from '../hooks/useUser'
-
-const Navbar = () => {
-
-    const { logout, user } = useUser();
-    console.log(user)
-
-  return (
-    <div className='border-b-[1px] border-neutral-800 p-5'>
-        <div className='flex justify-between'>
-            <div>
-                Lingosta
-            </div>
-            <div>
-                { !user && (
-                    <button onClick={() => logout()}>Logout</button>
-                )}                
-            </div>
-        </div>
-    </div>
-  )
-}
-
-export default Navbar
